refactor: Move get_query_dict to utils
Keep webutils with only API-related functions.
Also creates a separate testing module for the utils.
